test: use a common tray screenshot module entry point
